Move date formatters to intl package
