+++ packages/common/src/intl/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './IntlUtils';
+export * from './DateUtils';
