+++ packages/common/src/ui/components/domain/StatusCrumbs/StatusCrumbs.tsx
@@ -3,8 +3,7 @@ import Breadcrumbs from '@mui/material/Breadcrumbs';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import { ChevronDownIcon } from '@common/icons';
-import { useFormatDateTime } from '@common/utils';
-import { useTranslation } from '@common/intl';
+import { useTranslation, useFormatDateTime } from '@common/intl';
 import { VerticalStepper } from '../../steppers/VerticalStepper';
 import { PaperHoverPopover, PaperPopoverSection } from '../../popover';
 import { useIsSmallScreen } from '@common/hooks';
