+++ packages/common/src/ui/layout/tables/hooks/useColumns/useColumns.tsx
@@ -7,10 +7,8 @@ import {
   ColumnAlign,
   Column,
 } from '../../columns/types';
-import { useFormatNumber } from '@common/intl';
-import { useFormatDateTime } from '@common/utils';
+import { useFormatNumber, useFormatDateTime, DateUtils } from '@common/intl';
 import { BasicCell, BasicHeader } from '../../components';
-import { DateUtils } from '@common/utils';
 import { SortBy } from '@common/hooks';
 import { ColumnDefinitionSetBuilder, ColumnKey } from '../../utils';
