+++ packages/common/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './numbers';
 export * from './quantities';
 export * from './formatters';
 export * from './testing';
-export * from './dates';
 export * from './arrays';
 export * from './regex';
 export * from './pricing';
